requirement.txt
Django==1.11.1
MySQL-python==1.2.5
appdirs==1.4.3
django-auth-ldap==1.2.12
kafka-python==1.3.1
django-suit==0.2.25
packaging==16.8
pyparsing==2.2.0
python-ldap==2.4.38
pytz==2017.2
requests==2.14.2
six==1.10.0
uWSGI==2.0.15
wsgiref==0.1.2
xlrd==1.1.0
xlwt==1.3.0
python-memcached==1.58
crypto==1.4.1
pycrypto==2.6.1